requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langchain-ollama
chromadb==0.5.3
pypdf==4.3.1
docx2txt==0.8
streamlit==1.37.0
python-dotenv==1.0.1
pydantic==2.8.2